[#60] refactor: Use Case 분리
