+++ iOS/SideDish/Banchan/UseCases/BanchanDetailUseCase.swift
@@ -1,0 +1,9 @@
+//
+//  BanchanDetailUseCase.swift
+//  Banchan
+//
+//  Created by Chaewan Park on 2020/04/29.
+//  Copyright © 2020 Chaewan Park. All rights reserved.
+//
+
+import Foundation
