+++ iOS/SideDish/Banchan/Views/BanchanCell.swift
@@ -34,7 +34,7 @@ class BanchanCell: UITableViewCell {
         priceStackView.spacing = CGFloat(!(banchan?.isOnSale ?? true) ? 4 : 0)
         banchan?.badge?.forEach { badgeView.addBadge($0) }
         guard let url = banchan?.image else { return }
-        BanchanUseCase.performImageFetching(with: NetworkManager(), url: url) {
+        ImageUseCase.performImageFetching(with: NetworkManager(), url: url) {
             guard let image = UIImage(data: $0) else { return }
             DispatchQueue.main.async { self.banchanImageView.image = image }
         }
